04 auth

+++ controllers/contactsControllers.js
@@ -1,8 +1,9 @@
 import httpError from "../helpers/HttpError.js";
-import { Contact } from "../models/UseModels.js";
+import { Contact } from "../models/ContactsModels.js";
 async function getAllContacts(req, res, next) {
   try {
-    const allContacts = await Contact.find();
+    const owner = req.user._id;
+    const allContacts = await Contact.find(owner);
     if (!allContacts) {
       throw httpError(404);
     }
@@ -25,8 +26,9 @@ async function getOneContact(req, res, next) {
 }
 async function deleteContact(req, res, next) {
   try {
+    const owner = req.user._id;
     const { id } = req.params;
-    const allContacts = await Contact.findByIdAndDelete({ _id: id });
+    const allContacts = await Contact.findByIdAndDelete({ _id: id, owner });
     if (!id) {
       throw httpError(404);
     }
@@ -37,12 +39,14 @@ async function deleteContact(req, res, next) {
 }
 async function createContact(req, res, next) {
   try {
+    const owner = req.user._id;
     const { name, email, phone, favorite } = req.body;
     const createContact = await Contact.create({
       name,
       email,
       phone,
       favorite,
+      owner,
     });
     res.status(201).json(createContact);
   } catch (error) {
@@ -51,9 +55,13 @@ async function createContact(req, res, next) {
 }
 async function updateContact(req, res, next) {
   try {
+    const owner = req.user._id;
     const { id } = req.params;
     const editContacts = req.body;
-    const data = await Contact.findByIdAndUpdate({ _id: id }, editContacts);
+    const data = await Contact.findByIdAndUpdate(
+      { _id: id, owner },
+      editContacts
+    );
     res.status(200).json(data);
   } catch (error) {
     next(error);
@@ -61,10 +69,11 @@ async function updateContact(req, res, next) {
 }
 async function updateFavoriteContact(req, res, next) {
   try {
+    const owner = req.user._id;
     const { id } = req.params;
     const favorite = req.body;
     const data = await Contact.findByIdAndUpdate(
-      { _id: id },
+      { _id: id, owner },
       { favorite: favorite }
     );
     if (!data) {
@@ -83,3 +92,8 @@ export {
   updateContact,
   updateFavoriteContact,
 };
+// afwfawfawf12124124
+// {
+//   "email": "wfafwfawf@gmail.com",
+//   "password": "afwfafawf12124124"
+// }
